SGPR and SGPFITC now have `data` as part of the state tests using these models have been updated all _unsorted_tests (except for MC related) pass Change to float64 as default float required!

+++ gpflow/config/__config__.py
@@ -20,7 +20,7 @@ _ENV_FLOAT = "GPFLOW_FLOAT"
 _ENV_INT = "GPFLOW_INT"
 
 _JITTER_VALUE = 1e-6
-_FLOAT_VALUE = np.float32
+_FLOAT_VALUE = np.float64
 _INT_VALUE = np.int32
